Refactored mon_alert_list into alerting namespace in #95 PR #292
