+++ api/mon/urls.py
@@ -5,5 +5,6 @@ urlpatterns = patterns(
 
     url(r'^vm/', include('api.mon.vm.urls')),
     url(r'^node/', include('api.mon.node.urls')),
+    url(r'^alert/', include('api.mon.alerting.urls')),
     url(r'^', include('api.mon.base.urls')),
 )
