+++ api/mon/views.py
@@ -4,3 +4,5 @@ from api.mon.vm.views import *  # noqa: F401,F403
 from api.mon.node.views import *  # noqa: F401,F403
 # noinspection PyUnresolvedReferences
 from api.mon.base.views import *  # noqa: F401,F403
+# noinspection PyUnresolvedReferences
+from api.mon.alerting.views import *  # noqa: F401,F403
